fix(action-controller): keep LiveTestResponse + TestSession per Rails layout
Review against Rails source: `LiveTestResponse` and `TestSession` are
defined in `action_controller/test_case.rb` (lines 184, 197), not in
`action_dispatch/testing/test_response.rb` / `test_process.rb`. Those
two files define `ActionDispatch::TestResponse` and the `TestProcess`
module respectively — different surfaces entirely.

Revert: put LiveTestResponse / TestSession back in
action-controller/test-case.ts and delete the misplaced
test-response.ts / test-process.ts. TestRequest stays in
action-dispatch/testing/ since Rails defines
ActionDispatch::TestRequest there.

Implementing real TestResponse (parsed_body / from_response) and the
full TestProcess module is deferred — both add new behavior and are
better done in their own PR.
