+++ packages/actionpack/src/action-dispatch/testing/index.ts
@@ -1,4 +1,2 @@
 export { IntegrationTest, type IntegrationRequestOptions } from "./integration.js";
 export { TestRequest } from "./test-request.js";
-export { LiveTestResponse } from "./test-response.js";
-export { TestSession } from "./test-process.js";
